refactor fixture parse to use correct data
