+++ scrapr/json/lokFixtures.json
@@ -1,1 +1,1 @@
-[{"date":"Wednesday 26th April 2017","homeTeam":{"fullName":"Auchinleck Talbot","prettyName":"Auchinleck","homeGround":{"groundName":"Beechwood Park","groundTown":"Auchinleck","groundPostcode":"KA12"},"crestUrl":"https://static1.squarespace.com/static/552138ebe4b0abe69be47a2c/55b234a2e4b03d4ada444a70/55b234a2e4b03d4ada444aaa/1349466586068/1000w/10576389-20537545-thumbnail.jpg","twitter":"@steadythebuffs","website":"null"},"awayTeam":{"fullName":"Glenafton Athletic","prettyName":"Glenafton","homeGround":{"groundName":"Loch Park","groundTown":"New Cumnock","groundPostcode":"KA13"},"crestUrl":"https://static1.squarespace.com/static/552138ebe4b0abe69be47a2c/55b234a2e4b03d4ada444a70/55b234a2e4b03d4ada444aaa/1349466586068/1000w/10576389-20537545-thumbnail.jpg","twitter":"@steadythebuffs","website":"null"},"time":"18:45","unixTime":1493228700}]
+[{"date":"Wednesday 26th April 2017","homeTeam":{"fullName":"Auchinleck Talbot","prettyName":"Auchinleck","homeGround":{"groundName":"Beechwood Park","groundTown":"Auchinleck","groundPostcode":"KA12"},"crestUrl":"https://static1.squarespace.com/static/552138ebe4b0abe69be47a2c/55b234a2e4b03d4ada444a70/55b234a2e4b03d4ada444a7d/1349462491098/1000w/10576389-12551402-thumbnail.jpg","twitter":"@atfc1909","website":"null"},"awayTeam":{"fullName":"Glenafton Athletic","prettyName":"Glenafton","homeGround":{"groundName":"Loch Park","groundTown":"New Cumnock","groundPostcode":"KA13"},"crestUrl":"https://static1.squarespace.com/static/552138ebe4b0abe69be47a2c/55b234a2e4b03d4ada444a70/55b234a2e4b03d4ada444a98/1349461881293/1000w/10576389-15546177-thumbnail.jpg","twitter":"@flowgently","website":"null"},"time":"18:45","unixTime":1493228700}]
